Refactor alert endpoint handling by moving functionality to new endpoints package and updating initialization logic
